refactor(api): split 4 god-modules into cohesive sub-modules
Extract single-responsibility sub-modules from 4 monolithic files
without changing behavior or public API contracts:

1. admin.ts (327 lines):
   - admin-users.controller.ts: user listing, suspension, role management
   - admin-workers.controller.ts: worker moderation & bulk actions
   - admin-audit.controller.ts: audit log query handler
   - admin-stats.controller.ts: dashboard statistics metrics
   - admin.ts: facade re-exporting sub-module handlers

2. workers.ts (382 lines):
   - worker-crud.controller.ts: worker profile CRUD & pagination
   - worker-search.controller.ts: full-text & advanced search
   - worker-reputation.controller.ts: Stellar on-chain reputation
   - workers.ts: facade re-exporting sub-module handlers

3. worker.service.ts (567 lines):
   - worker-query.service.ts: cursor, geo-radius & offset queries
   - worker-crud.service.ts: worker creation, update, soft deletion
   - worker-media.service.ts: media attachment processing
   - worker.service.ts: facade re-exporting sub-module functions

4. auth.service.ts (382 lines):
   - auth-session.service.ts: login, JWT & refresh token rotation
   - auth-recovery.service.ts: registration & password reset flows
   - auth-2fa.service.ts: TOTP secret generation & verification
   - auth.service.ts: facade re-exporting sub-module functions

Verification:
- Preserved route wiring with zero changes to routes/.
- All 160 unit tests across 10 test suites passed.

Closes #1, #2, #3, #4
